requirements/common.txt
PyYAML>=6.0.1
bs4>=0.0.1
requests>=2.31.0
requests-cache==1.1.0
